+++ src/test/html/dom-module-name-or-is_test.ts
@@ -11,10 +11,10 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
+
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer} from 'polymer-analyzer';
-import {FSUrlLoader} from 'polymer-analyzer/lib/url-loader/fs-url-loader';
+import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
 
 import {DomModuleNameOrIs} from '../../html/dom-module-name-or-is';
 import {Linter} from '../../linter';